requirements.txt
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-cors-middleware==1.3.1
django-debug-toolbar==1.11
django-oauth-toolkit==1.2.0
django-rest-framework==0.1.0
djangorestframework==3.9.1
drf-tracking==1.5.0
drf-yasg==1.13.0
flex==6.13.2
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
kombu==4.3.0
MarkupSafe==1.1.0
mysql-connector-python==8.0.14
mysqlclient==1.4.1
oauthlib==3.0.0
pep8==1.7.1
protobuf==3.6.1
pytz==2018.9
PyYAML==4.2b4
requests==2.21.0
rfc3987==1.3.8
ruamel.yaml==0.15.86
sentry-sdk==0.7.2
six==1.12.0
sqlparse==0.2.4
strict-rfc3339==0.7
swagger-spec-validator==2.4.3
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
vine==1.2.0
